Rename script family.py to case.py
