+++ cg/cli/set/case.py
